Keylogger/requirements.txt.txt
pynput
Pillow
pyperclip